+++ code/lib/Display.cpp
@@ -10,12 +10,12 @@ private:
 public:
     Display() = default;
     virtual void printSimpleText(char text[]) {}
-    virtual void printData(char text1[], int data, char text2[]) {}
-    virtual void printData(char text1[], int data1, char text2[], int data2, char text3[]) {}
-    virtual void printIn(char text[], int x, int y) {}
-    virtual void printIn(int data, int x, int y) {}
+    virtual void printData(char text1[], int8_t data, char text2[]) {}
+    virtual void printData(char text1[], int8_t data1, char text2[], int8_t data2, char text3[]) {}
+    virtual void printIn(char text[], int8_t x, int8_t y) {}
+    virtual void printIn(int data, int8_t x, int8_t y) {}
     virtual void showClock(MyTime timeIn) {}
-    virtual void blinkAt(int x, int y) {}
+    virtual void blinkAt(int8_t x, int8_t y) {}
     virtual void noBlink() {}
     virtual void clockSym() {}
     virtual void noClockSym() {}
